Update calendar to improve reusability
